refactor TextView into graphics-server
