+++ services/gam/Cargo.toml
@@ -16,9 +16,6 @@ keyboard = { path = "../keyboard" }
 trng = { path = "../trng" }
 heapless = "0.5"
 log = "0.4"
-hash32 = "0.1.0"
-hash32-derive = "0.1.0"
-
 com = { path = "../com"}
 
 blitstr = { path = "../../../blitstr" }
