+++ services/gam/src/lib.rs
@@ -5,6 +5,7 @@ use api::*;
 
 use xous::ipc::*;
 use core::fmt::Write;
+use graphics_server::api::{TextOp, TextView};
 
 pub fn draw_textview(gam_cid: xous::CID, tv: &mut TextView) -> Result<(), xous::Error> {
     let mut sendable_tv = Sendable::new(tv).expect("can't create sendable textview");
